+++ tallerTrabajador/src/tallertrabajador/EstadoCivil.java
@@ -1,0 +1,9 @@
+package tallertrabajador;
+
+/**
+ *
+ * @author Adrian Verdugo
+ */
+public enum EstadoCivil {
+    CASADO, SOLTERO, DIVORCIADO, VIUDO;
+}