+++ trabajadores/src/trabajadores/Cargo.java
@@ -1,0 +1,9 @@
+package trabajadores;
+
+/**
+ *
+ * @author Adrian Verdugo
+ */
+public enum Cargo {
+    JEFEDEPTO, JEFEAREA, JEFESECCION, JEFEDIVICION
+}