+++ trabajadores/src/trabajadores/Rango.java
@@ -1,0 +1,9 @@
+package trabajadores;
+
+/**
+ *
+ * @author Adrian
+ */
+public enum Rango {
+    AYUDANTE, OPERARIO, PRIMERACLASE, EJECUTIVO, TRAINER
+}