+++ trabajadores/src/trabajadores/Seccion.java
@@ -1,0 +1,10 @@
+
+package trabajadores;
+
+/**
+ *
+ * @author Adrian
+ */
+public enum Seccion {
+    VENTAS, CONTABILIDAD, PRODUCCION, OPERACIONS, DISTRIBUICION
+}