+++ trabajadores/src/trabajadores/likidacion_sueldo.java
@@ -1,0 +1,10 @@
+package trabajadores;
+
+import java.util.LinkedList;
+
+public class likidacion_sueldo {
+
+    public void pago(crea_funcionarios lista) {
+
+    }
+}